More renames, move web handlers creation to hook
